requirements.txt
discord.py==2.3.2
openai==1.3.3
yfinance==0.2.31
pandas==2.1.1
matplotlib==3.8.1
Pillow==10.1.0
pytrends==4.9.2
beautifulsoup4==4.12.2
httpx==0.25.1
requests==2.31.0
lxml==4.9.3
numpy==1.26.1
